+++ ChangeLog.md
@@ -14,6 +14,8 @@
 * Fix a bug in which `normalizeDec` would report an incorrect number of
   `datatypeVars` for GADT declarations with explicit return kinds (such as
   `data Foo :: * -> * where`).
+* Fix a bug in which `freeVariablesWellScoped` would sometimes not preserve
+  the left-to-right ordering of `Name`s generated with `newName`.
 
 ## 0.2.10.0 -- 2018-12-20
 * Optimization: `quantifyType` now collapses consecutive `forall`s. For